delete:warning
As a user, I would like to be warned whenever deleting a position or person from the database.

+++ assets/js/people.js
@@ -217,8 +217,15 @@ $('#ppl-save-btn').click(function(){
 });
 
 $('#ppl-delete-btn').click(function(){
+    var ans = confirm("Delete this person from the database?")
+    if (answer) {
     console.log(getCurrentID());
     deletePerson(getCurrentID());
+    alert("Deleting the person:" + pplfname + " " + ppllname);
+    }
+    else {
+   //do nothing
+    }
 
  });
