+++ assets/js/positions.js
@@ -235,8 +235,15 @@ $('#save-btn').click(function(){
 
 
 $('#delete-btn').click(function(){
+    var answer = confirm("Delete this position from the database?")
+    if (answer) {
     console.log(getCurrentID());
     deletePosition(getCurrentID());
+    alert("Deleting the position:" + getCurrentID());
+    }
+    else {
+   //do nothing
+    }
 
  });
